pyproject.toml
[project]
name = "cursor-worker"
version = "0.1.0"
description = "FastAPI + Celery + Redis background jobs"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "celery[redis]>=5.4.0",
    "redis>=5.2.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.0",
    "sqlmodel>=0.0.22",
    "asyncpg>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "greenlet>=3.0.0",
    "httpx>=0.27.0",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["tasks*", "storage*", "models*", "clients*", "services*", "lib*"]

[tool.setuptools]
py-modules = ["config", "worker", "api"]

[tool.ruff]
line-length = 100
target-version = "py311"
